style: Move invalidation::element::collector to invalidation::element::state_and_attributes.
Since it's what that file is about.
