+++ components/style/invalidation/element/mod.rs
@@ -4,8 +4,8 @@
 
 //! Invalidation of element styles due to attribute or style changes.
 
-pub mod collector;
 pub mod element_wrapper;
 pub mod invalidation_map;
 pub mod invalidator;
 pub mod restyle_hints;
+pub mod state_and_attributes;
